requirements.txt
langchain>=0.0.200
openai>=0.27.0
chromadb>=0.3.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
pytest>=7.0.0
pydantic>=2.5.2
litellm>=1.10.2
ipykernel>=6.27.1
langchain-openai==0.0.2
boto3>=1.33.2,<1.34.35
aioboto3==12.3.0
